move config proxy to devices
