+++ src/components/devices/OpenwbConfigProxy.vue
