+++ src/views/HardwareInstallation.vue
@@ -254,7 +254,7 @@ library.add(fasPlus, fasNetworkWired, fasMicrochip, fasTrash);
 
 import ComponentStateMixin from "@/components/mixins/ComponentState.vue";
 
-import OpenwbConfigProxy from "@/components/OpenwbConfigProxy.vue";
+import OpenwbConfigProxy from "@/components/devices/OpenwbConfigProxy.vue";
 
 export default {
 	name: "OpenwbHardwareInstallation",
